Add support for getting appcast feed URL from sources.
As an alternative to win_sparkle_set_appcast_url(), it is possible to
specify the feed in win32 resources. This isn't as convenient as
Info.plist on OS X, but at least it still has the advantage of being
externally readable.

The resource has to be named "FeedURL" and have type set to "APPCAST".
The PSDK example now uses it.