Add Win64 or WOW64 to user agent string as appropriate.
Add "(Win64)" to the user agent string when running 64bit binary. Add
"(WOW64)" when running a 32bit one, but on 64bit Windows OS.

This is consistent with what most browsers do and it's useful for
determining what updates to serve. For example, a WOW64 installation may
get an update to convert it to natively 64bit version.

Fixes pull request #5 (with a different, more robust implementation).