Don't poll host app for termination
Instead of polling the host app for termination we now just display a
message to the user saying that we were unable to shutdown the
application at this time and presenting them with a retry button. If the
application isn't ready when the user presses the retry button nothing
happens. Otherwise we launch the installer and terminate the host.