Add API for silent on-demand update checks.
There might be cases where the application would want to check for
updates manually. This would override the users request to not perform
update checks, if that is enabled, so should be used with care.

Add in_sparkle_check_update_without_ui() for this purpose. It is not
completely UI-less, as it still shows the "update available" window, but
it doesn't show any UI unless there actually is an update.