Created a Full Integration mode
The new full integration calls allow WinSparkle to be called with zero
UI.  Callbacks are heavily used to make up for the lack of UI.  The
calling application is then responsible for implementing this callbacks
and giving feedback to the user.  This was developed to work with a game
that normally runs in full screen.