Make sure WinSparkle's windows are always toplevel
In a somewhat peculiar wx API choice, wxDialog with NULL parent is made
a child of the (main) toplevel window unless the wxDIALOG_NO_PARENT flag
is used. That was not the intent of this code, WinSparkle windows are
meant to be parent-less. Use wxDIALOG_NO_PARENT explicitly.

This wouldn't matter under normal conditions, because WinSparkle is
meant to show only one window, but it is possible to get it into a state
when it shows two during second launch of the host application: one will
be AskPermissionDialog. If the user doesn't dismiss it and explicitly
checks for updates, the regular UpdatesDialog will appear - and without
wxDIALOG_NO_PARENT, would become a child of modally-shown
AskPermissionDialog. When the users dismiss AskPermissionDialog
afterwards, WinSparkle's UI will be in inconsistent case, leading to
asserts and eventually failure to shut down the UI thread.

This fix is still not perfect: interaction with UpdatesDialog are lost
after the user dismisses AskPermissionDialog. This fix "only" makes this
corner case safe, but not ideal from UX point of view.

Fixes #44.