Add CMake build system for MinGW
WinSparkle can be used in MinGW applications. But can't be built with
MinGW. bkl files support only Visual Studio. CMake build system makes
it possible to build WinSparkle with MinGW. All CMake stuff placed in
cmake directory to interfere with the official build as little as
as possible.

Qt example can be build with Qt5 only.