Fix compilation in directory with spaces in name
Make sure to quote any absolute paths (expanded through variables) used
in WinSparkle.Targets so that the commands work correctly in directories
with whitespace in their names.

WinSparkle couldn't be built if the solution path has blanks in it. This
is the case, for example, if your WinSparkle project folder sits in the
Visual Studio default location for projects which happens to have
"Visual Studio 2013" in it, for example, if you are using Visual Studio
2013.

Fixes #90.