Fix WinSparkle.Targets
WinSparkle can't be built if the solution path has blanks in it. This
is the case, for example, if your WinSparkle project folder sits in the
Visual Studio default location for projects which happens to have
"Visual Studio 2013" in it, for example, if you are using Visual Studio
2013.

This commit encloses the solution path in double quotes to deal with
those blanks.

Fixes #90