Warn about use of insecure http:// URLs
Log warnings when user code attempts to use WinSparkle insecurely. Check
not only appcast feed, but also  download and release notes links if
present.

Do it as early as possible to make noticing it more likely, i.e. not
when actually used, but when set or fetched in the appcast (even if
there's no update to show at the time).

See #94.