Use WinInet asynchronously
This is more complicated, but necessary to be able to cancel download
operations in progress. With the (synchronous) InternetReadFile() call
previously used, the call was all-or-nothing and couldn't be cancelled
by the user if in progress. This would present a problem e.g. when
suddenly loosing connectivity, in which case the call would block
indefinitely waiting for data.

Instead, use the more complicated async API and check for user-initiated
cancellation while waiting for download data.

Fixes #108.