Fix broken OnEndElement in XML parsing
Due to easy to make typos in parts of the code, it wasn’t properly
closing <sparkle:version> elements. Fix by not only correcting the typo,
but organizing the code in a way that prevents such accidents.

Longer term, Expat needs to be replaced with a DOM-based XML API for
safer and more maintainable parsing.

Fixes #123.