Disable HTTP caching
Bypass locale WinInet cache in HTTP requests, always getting the latest
data. For appcast feeds, this was previously the case only for manual
checks, now is always done.

For downloads, this is a new thing. Also disable storing downloaded
data in WinInet cache, which is particularly useful in case of
downloads: there was no point in keeping a second copy in there.

This is consistent with Sparkle for macOS and makes WinSparkle work
better with misconfigured servers, with little negative impact on those
properly configured.

See #33 for real-life impact.