Fix update check not triggered
This applies to new installations only, where we expect the check to
happen immediately on the second run.

If the LastCheckTime registry value was not set, it would default to
current time, meaning that a check wouldn't happen until after the
interval passed. If the user always closes the app before reaching the
interval, then a check would never happen.

This simply reverts back to previous behavior. Give LastCheckTime a
default value of 0 so that a check happens immediately if there's
been no previous check.