Add NuGet package
Add a native package for C++ projects in Visual Studio. All that is
required to use it is to add the package with NuGet; include paths and
library paths are added to the project automatically.

Binaries for both x86 and x86_64 are included; the appropriate one is
used automatically.

To be perfectly portable, and support Windows XP, the binaries need to
be built with Visual Studio 2017.

Modelled after Microsoft.Web.WebView2 which does the same very nicely.