Add limited support for <sparkle:criticalUpdate> (#244)
Parse <sparkle:criticalUpdate> and don't allow skipping updates if
present. The code recognizes <sparkle:criticalUpdate/> both inside
<sparkle:tags> (Sparkle 1 way) and in <item> scope (Sparkle 2).

This does not parse the specific version string embedded in the
criticalUpdate tag.

The UI prevents the user from skipping the update or choosing to be
reminded later, but isn't outright user hostile and doesn't prevent them
from closing the window. That's because we don't know the circumstance -
the user may e.g. be doing something urgent at the moment.

Co-authored-by: Edwin Clement <edwinclement08@gmail.com>