Sanitize spurious whitespace in malformed URLs
Sanitize URLs from <link> and <sparkle:releaseNotesLink> elements so
that they don't contain leading/trailing whitespace. Such links are
invalid, but apparently exist in the wild ("for legibility"). Previously
sanitization was done by MSIE internally, so we never noticed; it is
exposed now by the WebView2 backend in a rather spectacular way: by
opening release notes in Edge if it's the default browser or showing
chrome-less window in the top left corner of the desktop if not.

WinSparkle shouldn't do crazy things like that in any case, and between
rejecting such appcast items as malformed and silently correcting them,
the latter is more pragmatic.

Fixes #252.