Only overwrite with suitable enclosure data
If just overwriting enclosure data multiple enclosure-tags may result
in wrong update information. Consider an update for multiple
platforms where the enclosure tag for windows is not the last tag.

This change only allows enclosure data to be overwritten if it
is suitable for window. It leaves the current behaviour intact by
accepting enclosure data if no Os is set.