Check Os in IsValid()
If a non-windows Os is set the update item should not be valid.
Consider an update item with an enclosure for a non-windows
version.

This change additionally checks if Os is either empty or starts
with OS_MARKER. It also moves the implementation into appcast.cpp
so it has access to OS_MARKER_LEN and OS_MARKER.