Recognize major.minor-build as OS version number too
As follow-up to da4e7e2b, understand major.minor-build syntax in
<sparkle:minimumSystemVersion> in addition to major.minor.build.

This is a forward compatiblity measure: by using this syntax, the feed
will remain compatible with versions prior to v0.8.2. The old version
will fail to scan a string using '-' and will only see major.minor
component - and so won't misinterpret the buildnumber as service pack
version.

Should only be used if compatibility with pre-0.8.2 WinSparkle is
a concern.