Fix base64_encode including '\0' in output.
CryptBinaryToStringA([in, out] pcchString) has unusual behavior

> If pszString is NULL, the function calculates the length of the return
> string (including the terminating null character) in TCHARs

> If pszString is not NULL and big enough, the function converts the
> binary data into a specified string format including the terminating null
> character, but pcchString receives the length in TCHARs,
> *not including the terminating null character*

So it requests a buffer size with room for a terminating '\0',
and will write that terminator, but won't count it in the final length.
And this std::string likewise implies a '\0'-termination,
but doesn't consider that part of std::string::size().

So the terminator written by CryptBinaryToStringA shouldn't be left
as part of the content of str.